Always return the service name in the catalog
Fix implemention of service catalog creation in v2, to handle the case
when the service name is not available. Using service.get('name', '')
does not work because that doesn't invoke the __getitem__() from
sql.DictBase parent class of Service. Instead the service name have to
be directly read from the 'extra' attribute of the service object using
service.extra.get('name', '').
The name attribute of the service will now always be included in the
service catalog. The value will be defaulted to empty string if no
service name is available.

Change-Id: I305abd8446fee57af18aebf7873e3c67a1b83c92
Closes-Bug: #1393518