Correct test_auth_unscoped_token_project for result ordering
The order of the roles in the authenticate result will vary depending
on how a query returns results, so the target role may not be the
first one. This can be seen by setting the PYTHONHASHSEED value
to something other than 0.

With this fix, a search is done rather than checking only the first
element so that the test works even when the order of the roles is
different.

Change-Id: I9ea13a2a7a242336d3658e6027d305f7371e7654
Partial-Bug: #1348818