Correct test_get_v3_catalog test for result ordering
The order of the endpoints and interfaces lists will vary depending on
how a query returns results, so the endpoints and interfaces may not
be in the same order as the expected dict. This can be seen by setting
the PYTHONHASHSEED value to something other than 0.

With this fix, searches are done rather than using the built-in dict
equal method.

Change-Id: I4623401e93c4dd2e9347eb485dc701891af79a44
Partial-Bug: #1348818