Correct version tests for result ordering
The order of elements in the versions response depends on how the dict
is ordered which may differ. This can be seen by setting the
PYTHONHASHSEED value to something other than 0.

With this fix, code is written that compares the version elements so
that the order as returned doesn't matter.

Change-Id: I24f7fd2fb5e5fc6518c284ce728d2a84e1549272
Partial-Bug: #1348818