Cleanup eventlet use in tests
The tests in test_wsgi should stop the server once the test is over,
otherwise it would continue running and potentially affect other
tests.

Also, there were a couple of tests that did
environment.use_eventlet() in setUp, which has no effect since this
was already done in keystone.tests.core, and it's incorrect to put
this in a test setUp since it would affect other tests as it's an
application-wide library setting.

Closes-Bug: #1400565
Change-Id: Ibb00a1afa64cf46aaa0eafaca895bcf7a1f69eb0