Fix update role without name using LDAP
When the keystone server was configured to use the LDAP assignment
backend and a role was updated and no name was included in the update
dict the operation would fail with a 500 Internal Server Error.

With this fix, an update without a name works.

(cherry picked from commit 93ab08a46bbf5d466b1a12fe9691565ec10b1179)

Change-Id: I234a36301a54ad4796cd421c42e85b15e9415902
Closes-Bug: #1241134