move region and service exist checks into manager layer
When creating or updating endpoint we should check if the region or
service passed exists. But it is confused in our current implementation:
some checks are in backends, some checks are in controller, some
checks are duplicated, some checks are missing.
Now, move these checks into manager where they should be.

Change-Id: I191394540ffe43f0ee9f0d28e00e70d44bfcf701