Make unit tests call the new, split out, role manager
This is the part of the more comprehensive split of
assignments, which rationalizes both the backend and controllers.
In order to make this change easier for reviewers, it is divided
into a number of smaller patches. Previous patches have:

- Moved role management into its own manager and drivers
- Fixed incorrect doc strings for grant driver methods
- Updated controllers to call the new role manager

This patch updates any tests that make calls to manager role to
call the new role manager.

Future patches will:
- Refactor the assignment manager and drivers enabling
  projects/domains to be split out
- Split projects and domains into their own backend
- Split the controllers so they call the correct manager
- Update the tests to call the new correct manager

Partially implements: bp pluggable-assignments

Change-Id: I0a7a648e85bdce4e47c6eae47dcc40a438a5f0a5