Add support for data-driven backend assignment testing
Most backend assignment listing is planned to become based on
list_role_assignments. As such this method needs to be tested with
a myriad of different options. We want to avoid having acres of
test code or using complex test algorithms (which often mirror
the code under test) making it difficult to easily determine what
exactly is being tested. Ideally what you want is the ability to
write some kind of data driver shorthand for these tests, which
allows testers and reviewers to see "at a glance" if a test scenario
is appropriate and comprehensive.

This patch implements the basic code for a TestHelper that provides
the above by allowing a test to simply be a dict-based test plan.
A few examples of its use are also included. Subsequent patches will

- Add support for effective and inherited options
- Replace existing assignment tests with data-driven test plans and
  add additional tests to widen our coverage.

Partially Implements: bp data-driven-tests
Change-Id: I7023f9ea28251f7aa16e789b256422b401c42c7f