Add support for effective & inherited mode in data driven tests
Support the testing of inherited role assignments as well as
listing in effective mode to the data driven tests. The first
examples that use these capabilities are also included.

Follow on patches will extend the use of this to other existing
tests, as well as to widen coverage.

Partially Implements: bp data-driven-tests
Change-Id: I676c3889b8f7216e56b9e5982463fa5d5d6aa454