Add a check to see if a federation token is being used for v2 auth
If a user has a federation environment, and doesn't update nova
to use v3 auth in auth_token, then they will automatically be
using v2 auth, which causes a fairly useless error to bubble up.
We should check if the user is using a fedration token in v2 and
provider a better error message.

Change-Id: I9c490df8dd38e9fa1f393ebfa1296b64fb0c9527
Closes-Bug: #1419114