Move existing tests to unit
The existing test files are all moved under keystone.tests.unit,
except the existing keystone.tests.unit are left in place.

The .testr.conf is updated so that unit tests are run by default
in tox envs, and a tox env can override the tests to run by
setting OS_TEST_PATH.

This is so functional tests can sit in keystone.tests.functional.

Change-Id: I065d3f56e22f344abdadd92b3b384b002b02d989