Improved policy setting in the 'v3 filter' tests
The test case class IdentityTestListLimitCase (test_v3_filters.py)
always sets multiple policies for the each test defined in that class.
However, only one policy is really needed for each test case.
That appears to lead to unreliable policy checking.
It would be better to set only one policy required for the particular
test case (policy with the type of entity used in the test case).

Change-Id: I7892950b43c56ecce03677d8d592e71359bedf27
Closes-Bug: #1271273