Make RuleProcessor._UserType class public
Since keystone.auth.plugins.mapped.Mapped auth plugin must be aware what
kind of user is being authorized (ephemeral or local) we need to make
class RuleProcessor._UserType public. This means the class is going to be
renamed to UserType and taken out of the RuleProcessor's namespace.

Change-Id: I9a3c64448432190fc818ecdc79c429c57e09e476
Partially-Implements: bp federated-direct-user-mapping