Add in non-decorator notifiers
Add non-decorator based notifiers in. These notifiers will be used
in lieu of the decorator based notifiers so that the initiator can
be passed through to the manager for notification that does not
result in suboptimal design patterns where an argument is passed to
the manager for the sole purpose of the decorator inspecting for it
to populate data in the CADF notifications.

partially implements bp: cadf-everywhere

Change-Id: I684cf11325ffa7e6320088bd8b8ebb6cda29ad12