Rename test_content_types
This module used to do functional testing to ensure identical responses
for both JSON and XML, but since we have dropped support for XML, the
name of the module doesn't make much sense anymore.

What is left in this module is effectively the v2 parallel of test_v3,
so it should be renamed accordingly to make it easier for new
contributors to find it.

Change-Id: I63aaeacabfc998a3fff88c6959f50376ba960e94