Add domain_id checking in create_project
It was possible to create a project with invalid domain_id if
we are using engines that do not support foreign keys (such
as sqlite). If the engine being used supports foreign keys
the returned error was a generic Internal Server Error.

This patch adds a check in the create_project() method in
order to verify if the domain_id is valid (there is a domain
with such id), returning DomainNotFound error if it is case.

Co-Authored-By: Hugo Nicodemos <hugonicodemos@gmail.com>
Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Erickson Santos <erickson@lsd.ufcg.edu.br>

Change-Id: I8f394f46ab915844228de00c5a9223c1e3fb6be1
Closes-Bug: 1426496