Prevent calling waitall() inside a GreenPool's greenthread
Keystone eventlet wsgi server is started in a green thread from
the same green pool that is passed to the server itself
to be used for spawning client green threads.

To avoid it, it is reasonable to use different pools for spawning
a wsgi server and for its internal usage.

Change-Id: I1246e2b730a522480ed495fbb53ae376ee7593d5
Closes-Bug: #1423250