Stop debug logging of Ldap while running unit tests
Debug logging of Ldap generates at least 25Mb of log during
a full run of unit tests. Combine this with other logging and
we are close to blowing the 50Mb limit for a unit test run,
meaning we cannot currently, for instance, add new Ldap tests.

This patch changes the log level to INFO for Ldap.

Change-Id: I36265212d402c96c33435fb0ff5c73acbcb42c89
Closes-bug: 1427437