Move install of cryptography before six
With Ibca4b1765d06f239df113aa3ec367e60de61a225 a dependency on
cryptography was introduced.

This is causing issues for the centos7 devstack jobs.  cryptography
attempting to be installed before it's dependency "six".

I have filed an upstream issue with pip about this [1].  In the
mean-time, re-ordering the dependency should allow things to install
correctly.

This probably doesn't happen on Ubuntu because it ships with a version
of six that allows the cryptography package to at least install.

[1] https://github.com/pypa/pip/issues/2473

Change-Id: I0308aca7ac4d59afbec342ec387be088f7d360ca