'Assignment' has no attr 'get_domain_by_name'
The domain API has moved to the resource manager. While there are
redirects at the manager level so that most calls won't notice,
keystone-manage mapping_purge is making a direct driver call (bypassing
the manager). It needs to be updated to point to the resource manager
rather than the assignment manager.

Change-Id: Ia4fc89a0496bf9dc45300dbbe4cb776210340de4
Closes-Bug: 1427379