Refactor: make Fernet token creation/validation API agnostic
A lot of the Fernet token create / validate logic ended up in the v3
token provider methods, but there's nothing really API-specific about
it, so it should be moved somewhere such that it's available for use by
the v2 token provider code, which will just utilize a subset of the
available functionality (no domain scoped tokens, etc).

As a result of this change, everything about the implementation details
of Fernet tokens end up in the token_formatters module.

Change-Id: I6fe0fb581f156d14fc6be3772436cc03d909c4a6