Address nits for default cache time more explicit
This addresses minor nits in naming of functions when the cache
time was made more explicit in code via refactor. Notably the
`should_cache_fn` on the decorator now is called `should_cache` and
the `expiration_time_fn` is called `get_expiration_time`. The
declairation of the memoizing decorator in keystone.token.provider
has been made consistent with other uses by utilizing the `section`
kwarg instead of passing it as positional.

Change-Id: I19c50f1fa1fff56f04944bedf58c9438e4115322