Fix deprecated group for eventlet_server options
Several eventlet_server options were moved from the DEFAULT group
and the old DEFAULT group values are deprecated, but since the
deprecated_group wasn't set the oslo.config code was using the
options' new group as the deprecated group. This caused the
server to report that the new values were deprecated when used.

Change-Id: I98bb59b20e50490bdf054df68b5de055abc3db81
Closes-Bug: 1431088