Lookup identity provider by remote_id for websso
When using the websso feature in keystone, the identity provider is
looked up based on the value of the 'remote_id_attribute' environment
variable provided by the SAML assertion (or claim in the case of OpenID
Connect). Logic would dictate that the 'remote_id_attribute' value is
searched for against the 'remote_id' field in the backend where identity
providers are stored. Instead, we are doing the lookup against the 'id'
field.

This patch makes the websso code look the identity provider up by the
remote ID.

Change-Id: I1dc666782929902a3a4f478baef2b729757b0dc5
Closes-Bug: 1434701