Skip SSL tests because some platforms do not enable SSLv3
For "damned good security reasons" (POODLE) some platforms have
patched out SSLv3 from OpenSSL. Because Evenetlet cannot be configured
for specific versions (or ciphers) and httplib is extremely limited,
the tests for SSL have been explicitly skipped. These tests should
be potentially re-enabled in the functional suite for SSL terminated
endpoints.

This comes back to "do not terminate SSL in the eventlet wsgi" that
was determined in bug 1381365.

Change-Id: Ic4b446ceee9034de5b6530c2d79d798a903fcbbf
Related-Bug: #1381365
Closes-Bug: #1435174