Add openstack_user_domain to assertion
Currently, a keystone IdP does not provide the domain of the user
when generating SAML assertions. Since it is possible to have two
users with the same username but in different domains, this patch
adds an additional attribute called "openstack_user_domain"
in the assertion to identify the domain of the user.

Closes-Bug: 1442787
bp assertion-extra-attributes

Change-Id: I65d5c02c0a21f4d4c1b54f8aa56e27950d20badd