Adds proper isolation to templated catalog tests
Previously the templated catalog implementation wasn't really being
tested properly. The read/write tests, while passing, tested the KVS
implementation that will soon be removed.

A bad side effect of using KVS as a base is that the read operations
don't use the data from the template. They use only the data that
was modified by crud. This is why we didn't realize that the read
operations were incorrect.

Change-Id: I91ec863eabbf6074590e7f2ca43307b8b5133cf4