Manager support for projects acting as domains
Provide the Manager underpinnings for allowing a root project to act
as a domain. Subsequent follow-on patches will build on this to allow
the creation of such projects acting as domains at controller level,
provide full domain operations on such projects and the creation of
sub-projects that act as domains.

In this patch we provide manager support for creating a project
specifying a parent_id but not a domain_id. In the
controller layer we still normalize the domain_id if no
parent_id or domain_id were provided.

Co-Authored-By: Raildo Mascena <raildo@lsd.ufcg.edu.br>
Co-Authored-By: Rodrigo Duarte <rodrigods@lsd.ufcg.edu.br>

Change-Id: I21a89672dde27d9b9b34431e66235aac890e160e
Partially-Implements: bp reseller