Ignore multiple imports per line for six.moves
In the transition to Python3 we'll be heavily using six.move to import
builtins that have changed and modules that have moved. Having them all
on one line will, in many cases, make the code look cleaner. There is
exactly what we already do with the i18n translation helpers.

bp python3
Change-Id: I5c56a1cf36749d9f190e4360d5a12616380b6598