Use single connection in get_all function
As the number of LDAP users increase, any keystone commands that
need to look through all of the LDAP users tend to slow down.
After some investigation, I figured out that the cause of this
is making a new connection for each user in get_all.

In get_all, it calls _get_enabled function for each LDAP user.
In _get_enabled function, a new connection from get_connection()
is created, an then "enabled" value for a user is obtained.
Therefore, the more LDAP users you have, the slower the response
will be.

If a connection is made before the for loop in get_all, get all
"enabled" values from all users, and then close the connection,
then the response from keystone commands would be faster.

This code change was tested on a VM, and it actually speeded up
many keystone commands (user-list, user-role-add, user-delete,
user-get) from 50 seconds to 4 seconds.

Closes-Bug: #1451910
Change-Id: I3f5131c9eefa066f0d8ee5127322d686e09308df