Make sure LDAP filter is constructed correctly
This fixes an issue where, when querying Keystone via the v3 API, you
would get back an invalid LDAP filter, because None was coerced to the
string "None" and inserted into the middle of the query.

Conflicts:
        keystone/tests/unit/common/test_ldap.py

The conflict is due to imports being reorganized in the same area as
this change.

Change-Id: I9d45a4dca265b69e261f134118bb30c8cd128166
Closes-Bug: 1454309
(cherry picked from commit 2c6db4a3bb9e1718744b0e5b03af050fd2866182)