Don't assume project IDs are UUID format
Since Keystone has the ability to be deployed with an LDAP backend, we can't
guarantee that project IDs will always be UUID formatted. That change makes it
so that we attempt to convert project IDs to bytes before packing and if we
fail, pass the original project ID into the payload at the expense of a
slightly longer payload.

Change-Id: Id81ed23879ee7a9adeb50454ef0cb7acd13f1a0a
Related-Bug: 1459382
(cherry picked from commit f5f1caf9e9e4dc0d916fc48b32a408f477a87829)