Move endpoint_filter migrations into keystone core
This is the most basic way to move the endpoint_filter migrations into
the keystone core migrations path. Check to see if the migration was
performed already via the extension and if so skip performing it again.

DocImpact: Do not run the endpoint filter extension migrations, they are
now included in core and run by default.
Co-Authored-By: Jamie Lennox <jamielennox@redhat.com>
Partially implements: bp move-extensions

Change-Id: I51e711ed208dcc79531ac3a42b817063a2c712b8