Relax newly imposed sql driver restriction for domain config
The first implementation of the domain config API introduced a
new restriction in terms of which identity drivers could be SQL
backend (to avoid potential multi-process race conditions). This
restriction specified that only the default domain could be SQL
backend.

This patch resolves these issues and hence ensures the restrictions
in terms of SQL identity drivers are the same as the original config
file based approach (i.e. only one driver can be SQL backend, but
it can be for any given domain).

Although this patch changes the V8 driver interface, this is for
the domain_config backend which is still marked as experimental.

Closes-Bug: #1450344
Change-Id: Id1b4415cb5f263dc7afd9e25bd0da2b667f1c90a