Consolidate the fernet provider validate_v3_token()
When the Fernet token provider was implemented, it extended the
BaseProvider class. It also overrode most of the common methods
the BaseProvider implemented. Other token providers in Keystone (like the UUID
and PKI providers) just implement a _get_token_id method because token ids may
be different across providers.

This commit removes the validate_v3_token() method from the fernet.Provider.
This ensures that the Fernet provider uses the same validation method that all
the other token providers use.

Change-Id: Ib86dbe2229e8ebf2639cf2e52751554113c047a2