AuthContextMiddleware admin token handling
The AuthContextMiddleware shouldn't be re-implementing the
AdminTokenAuthMiddleware but using the request environment
context setting that the admin token middleware sets up.

This makes it so that admin token handling is in one place
rather than duplicating it and allows for an alternative
implementation of the admin token middleware.

The old behavior is left in place as deprecated to be removed
in a future release.

ReleaseNotesImpact
 - The paste.ini file is changed to put the admin token
   middleware first so that the auth context middleware can use
   the results. It's deprecated to have the admin token
   middleware after the auth context middleware.

Closes-Bug: 1473553
Change-Id: I658213699ac4af0abd08f893d9cf18ef0af5827d