Remove fileutils from oslo-incubator
The only function being used from oslo-incubator is
write_to_tempfile, which is being used in idp.py.

fileutils was recently graduated to oslo.policy and oslo.utils,
with file_open() and write_to_tempfile() NOT graduated.

We should include the small function in our own code, since
keystone is the only project using the function.

Change-Id: Ie5731a8f7b393ed8d6025ddf7d722c7688a42b3d