Fix log message in one of the v3 create call methods.
This error message is encountered when trying to get the domain_id for
a v3 create call (e.g, create a user using v3 API). The message is
misleading since it says, 'Invalid token found while getting domain ID
for list request' although the call is actually for 'create request'.

Change-Id: I091d172e21e96a77e2c3416aebb02791cabc15ac