Register fatal_deprecations before use
The keystone.tests.unit.test_backend_sql.DeprecatedDecorators
tests would fail when run individually with an error

 oslo_config.cfg.NoSuchOptError: no such option: fatal_deprecations

This is because the way that fatal_deprecations gets registered is
when versionutils.report_deprecated_feature() is called.

The fix is to call versionutils.report_deprecated_feature() before
trying to set the fatal_deprecations option.

Change-Id: I6ad2dd0a47f00ea2ba05d587e0291230e789f057
Closes-Bug: 1474069