Better error message when unable to map user
While authenticating federated users the mapping rules and REMOTE_USER
environment variables are used for setting user's id and name. In case
no sufficient information is provided (i.e. missing rule) an HTTP 401
Unauthorized error is returned and the error body is vague "Could not
map user". This is not very helpful for inexperienced users and does not
give them enough information what really went wrong. This patch changes
error message so it's more informative and gives clues on possible
reasons of problem.

Change-Id: I0504f5e31e9625fae92d71edf6d83d71d5ef4d36