Add groups in scoped federated tokens
Groups should be added also to project/domain federated tokens.
Rationale behind this is that federated token is an 'identity' of the
ephemeral user and there is no other way to retrieve required
information rather then directly from the token.

We also do a simple refactoring for a better readibility:

dict[x] = obj is used instead of dict.update({a:b})

Change-Id: I0b11d93836e5737d4d8fdfb5ec0351072eae7329
Implements: bp groups-in-federated-scoped-tokens