Fernet 'expires' value loses 'ms' after validation
When we create a Fernet token we get a token with an expiration
time which contains microseconds. When we validate a Fernet token,
the expiration time changes and comes back missing the microseconds.
We need to make the expiration time the same between creation and
validation time, having microseconds in both cases. This applies to
both v2 and v3 API.

Conflicts:
    keystone/tests/unit/test_v3_auth.py
    keystone/tests/unit/token/test_fernet_provider.py
    keystone/token/providers/fernet/core.py
    keystone/token/providers/fernet/token_formatters.py

NOTE: This backport differs from the patch to master to handle the
difference in oslo.timeutils versus the move to keystone's own
implementation of isotime in master, and to move the fix in
keystone/token/providers/fernet/core.py to avoid having to backport a
significant refactor Fernet's implementation.

Closes-Bug: #1459790
Change-Id: Ic6adbc2d69af03519a2c4ef66d558bcff9022049
(cherry picked from commit 9391f9307e834f437c6712797c903921881f2977)