Expose exception due to missing id of LDAP entity
There are cases when users get created in AD and don't always have
`user_id_attribute` field populated because it is only a requirement for
Openstack in the environment. When a user does not have a populated
`user_id_attribute`, all keystone queries fail, not just queries or rows
containing that user.

Show that the exception occurs and get_all fails.

Change-Id: Iac623f3a86d5c55d82d386048b1f908f59193e42
Partial-Bug: 1478579