Prevent exception due to missing id of LDAP entity
There are cases when users get created in AD and don't always have
`user_id_attribute` field populated because it is only a requirement for
Openstack in the environment. When a user does not have a populated
`user_id_attribute`, all keystone queries fail, not just queries or rows
containing that user.

Fix the query to LDAP to fetch only entries that have id assigned.

Change-Id: I0c7de36808ca6081ffc55a188a4583ddb5712bee
Closes-Bug: 1478579