Add user domain info to federated fernet tokens
When using the fernet token provider the domain info for federated
users was lost but this info is required (for rbac checks for
example).

The domain info for the federated user is part of the token data.
In the case of federated tokens, the domain id and name are the
values set in the [federation].federated_domain_name config option.

Closes-Bug: 1500459
Change-Id: I01e322e6f3567f42d08e2db952be4d58ee43b4b5