Show v3 endpoints in v2 endpoint list
v3 endpoints that are not migrated from v2 endpoints, were not shown in
v2 endpoint list. While it was more clear, separating v2 and v3
endpoints, it was not easy for tools or services which are still using
v2 APIs to get the endpoints created by v3 API.
Since it won't do harm to return more information, this commit is to
show v3 endpoints in API call to list v2 endpoints.
NOTE: public interface is required in a v2 endpoints, so only v3
endpoints with public interface are added into the list.

Change-Id: Id9b3822f87e973b6eb7dd9fc715c2dba76b9fe04
Closes-Bug: #1480270