Rationalize list role assignment routing
The v3 role assignment controller was created as a standard
controller even though it only supports the list method. This means
not only do we have code stubs that are unused, but this will
complicate the addition of separate entry points for listing role
assignments for a tree in the follow on patch.

This patch, therfore, rationalizes the routes so that this
is set up for just the list method, and removes the unused code
stubs.

Implements: bp list-assignment-subtree
Change-Id: I0c03accecdcf1608b50ad6b130c8a49d8a9b5b42