Update federation router with missing call
We can do federated authentication for specific identity providers with both
GET and POST operations, according to the code in
keystone/contrib/federation/routers.py. The inline documentation in the module
only said that GET was possible and didn't include documentation on the POST
operation. This commit updates the inline documentation to be consistent with
the code.

Change-Id: I88e8ab231957ba977c7d2fa7f8ee0443842ef584