Allow the PBR_VERSION env to pass through tox
The PBR_VERSION environment variables is useful to skip the auto
versioning that PBR does from git. This can be a problem downstream as we
want to hardcode the version number in a package and not count the
patches shipped on top. [1]

By default tox >= 2.0 strips everything out of the environment and so
this value is ignored so add it to the permitted environment variables
that are available in testing.

[1] http://docs.openstack.org/developer/pbr/packagers.html#versioning
[2] https://testrun.org/tox/latest/example/basic.html#passing-down-environment-variables

Change-Id: Ibf5429f6db0bfec77e393075e38203840bcd7c1e