Handle 16-char non-uuid user IDs in payload
If a user_id just happens to be of 16 character length, this will
cause the convert_uuid_bytes_to_hex function to improperly return
a UUID value instead of the user_id string unconverted.

This patch modifies the payload to indicate whether the ID was in
fact a UUID and the attempt to convert to bytes was successful.

This change has effect on more than just user IDs.  It also resovles
potential issues with project IDs, group IDs, IDP IDs, and scope IDs.

Change-Id: Ia4a4f760d67d8bbc22759c48fc800aef016b84ed
Closes-Bug: #1497461