Use deepcopy of mapping fixtures in tests
dict.copy() only performs a shallow copy of a dictionary. So by
modifying it after a copy you're still modifying the original dict.
Change this to do a deep copy so we don't mutate the fixtures.

I had not come across anywhere where this caused a problem, just
something I saw.

Change-Id: I9a1284a9bab2a9d0be74abac9931c4a81f583d6d