Perform middleware tests with webtest
Currently the middleware tests are performed by creating a fake request
object and then running the perform_request method. This doesn't take
into account if we were to use a more complex request class than the one
that is stubbed out.

By using webtest to make these requests we allow them to perform the
full webob traversal that they would in a real web request situation.

Change-Id: I37c0b48a8d564dcd4aee59b86e50f374f5b60531