Common arguments for fernet payloads assembly
Large sets of if/else statements are an indication that
object-oriented programming should be used instead.

As a first step towards making the fernet payload classes more
object-oriented, this change normalizes the arguments to the
"assemble" method so that create_token() doesn't have to know
details about the payload class.

Since all the methods work the same, the subclass implementations
don't need their own docstrings.

Change-Id: Id71e5cd91496e66849615f6f1ecdb1e5ad1129b2