De-duplicate fernet payload tests
There was a lot of copy-pasted code in the fernet payload tests.
Since all these tests are trying to assert the same thing (that the
disassembled data is the same as the assembled data) for different
payload types, the basic test should be in a function that can be
called for different payload types.

Change-Id: Ie7589e204ddd0266e51c54508528b7e5c9e1a850