Rename fernet methods to match expiration timestamp
Fernet tokens carry the token creation time as part of the Fernet
specification, which is an integer. This is not the case with the token
expiration time, which is carried in the payload of the Fernet token. Keystone
converts the timestamp to a floating point value to save space in the token.
Previously, all the methods handling this conversion, along with the related
tests, communicated that this conversion was to an integer instead of a
float. This change corrects the methods signatures in the Fernet token provider
to be consistent with the timestamp type that is used in the implementation.

Change-Id: Ibea9fa0d546948501ebcd9f9f826b1f1319246f2